APPT-998 - Removing bulk import feature flag
