+++ features/int.feature.flags.json
@@ -2,7 +2,6 @@
   "FeatureManagement": {
     "OktaEnabled": true,
     "JointBookings": false,
-    "BulkImport": true,
     "MultipleServices": true,
     "SiteSummaryReport": false
   }
