+++ src/api/Nhs.Appointments.Api/local.feature.flags.json
@@ -1,6 +1,5 @@
 {
   "FeatureManagement": {
-    "BulkImport": false,
     "JointBookings": false,
     "OktaEnabled": false,
     "TestFeatureEnabled": true,
