+++ src/api/Nhs.Appointments.Core/Features/Flags.cs
@@ -2,7 +2,6 @@ namespace Nhs.Appointments.Core.Features;
 
 public static class Flags
 {
-    public const string BulkImport = "BulkImport";
     public const string OktaEnabled = "OktaEnabled";
     public const string JointBookings = "JointBookings";
     public const string MultipleServices = "MultipleServices";
