--- tests/Nhs.Appointments.Api.Integration/BulkImportSerialToggleCollection.cs
@@ -1,8 +1,0 @@
-using Xunit;
-
-namespace Nhs.Appointments.Api.Integration;
-
-[CollectionDefinition("BulkImportSerialToggle")]
-public class BulkImportSerialToggleCollection : ICollectionFixture<object>
-{
-}
