+++ tests/Nhs.Appointments.Api.Integration/Scenarios/BulkImport/BaseBulkImportFeatureSteps.cs
@@ -13,42 +13,9 @@ using System.Threading.Tasks;
 using Xunit.Gherkin.Quick;
 
 namespace Nhs.Appointments.Api.Integration.Scenarios.BulkImport;
-public abstract class BaseBulkImportFeatureSteps(string flag, bool enabled) : FeatureToggledSteps(flag, enabled)
+public abstract class BaseBulkImportFeatureSteps : BaseFeatureSteps
 {
-    private HttpResponseMessage Response { get; set; }
-
-    [When("I import the following users")]
-    public async Task ImportUsers(DataTable dataTable)
-    {
-        const string usersHeader = "User,FirstName,LastName,appointment-manager,availability-manager,site-details-manager,user-manager,Region,Site";
-
-        var csv = BuildInputCsv(dataTable, usersHeader);
-
-        using var content = new MultipartFormDataContent();
-        using var fileContent = new ByteArrayContent(csv);
-
-        fileContent.Headers.ContentType = new MediaTypeHeaderValue("text/csv");
-        content.Add(fileContent, "file", "user-upload.csv");
-
-        Response = await Http.PostAsync("http://localhost:7071/api/user/import", content);
-    }
-
-    [When("I import the following site")]
-    public async Task ImportSite(DataTable dataTable)
-    {
-        const string sitesHeader =
-            "OdsCode,Name,Address,PhoneNumber,Longitude,Latitude,ICB,Region,Site type,accessible_toilet,braille_translation_service,disabled_car_parking,car_parking,induction_loop,sign_language_service,step_free_access,text_relay,wheelchair_access,Id";
-
-        var csv = BuildInputCsv(dataTable, sitesHeader);
-
-        using var content = new MultipartFormDataContent();
-        using var fileContent = new ByteArrayContent(csv);
-
-        fileContent.Headers.ContentType = new MediaTypeHeaderValue("text/csv");
-        content.Add(fileContent, "file", "user-upload.csv");
-
-        Response = await Http.PostAsync("http://localhost:7071/api/site/import", content);
-    }
+    protected HttpResponseMessage Response { get; set; }
 
     [Then("I receive a report that the import was successful")]
     public async Task AssertSuccessfulBulkImport()
@@ -63,7 +30,7 @@ public abstract class BaseBulkImportFeatureSteps(string flag, bool enabled) : Fe
     [Then(@"the call should fail with (\d*)")]
     public void AssertFailureCode(int statusCode) => Response.StatusCode.Should().Be((HttpStatusCode)statusCode);
 
-    private byte[] BuildInputCsv(DataTable dataTable, string headers)
+    protected byte[] BuildInputCsv(DataTable dataTable, string headers)
     {
         var sb = new StringBuilder(headers);
