+++ tests/Nhs.Appointments.Api.Integration/Scenarios/BulkImport/UserBulkImportFeatureSteps.cs
@@ -1,13 +1,27 @@
-using Nhs.Appointments.Core.Features;
-using Xunit;
+using Gherkin.Ast;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Threading.Tasks;
 using Xunit.Gherkin.Quick;
 
 namespace Nhs.Appointments.Api.Integration.Scenarios.BulkImport;
 
-[Collection("BulkImportSerialToggle")]
-[FeatureFile("./Scenarios/BulkImport/UserBulkImport_Disabled.feature")]
-public class UserBulkImport_BulkImportDisabled() : BaseBulkImportFeatureSteps(Flags.BulkImport, false);
+[FeatureFile("./Scenarios/BulkImport/UserBulkImport.feature")]
+public class UserBulkImportFeatureSteps : BaseBulkImportFeatureSteps
+{
+    [When("I import the following users")]
+    public async Task ImportUsers(DataTable dataTable)
+    {
+        const string usersHeader = "User,FirstName,LastName,appointment-manager,availability-manager,site-details-manager,user-manager,Region,Site";
 
-[Collection("BulkImportSerialToggle")]
-[FeatureFile("./Scenarios/BulkImport/UserBulkImport_Enabled.feature")]
-public class UserBulkImport_BulkImportEnabled() : BaseBulkImportFeatureSteps(Flags.BulkImport, true);
+        var csv = BuildInputCsv(dataTable, usersHeader);
+
+        using var content = new MultipartFormDataContent();
+        using var fileContent = new ByteArrayContent(csv);
+
+        fileContent.Headers.ContentType = new MediaTypeHeaderValue("text/csv");
+        content.Add(fileContent, "file", "user-upload.csv");
+
+        Response = await Http.PostAsync("http://localhost:7071/api/user/import", content);
+    }
+}
